+++ app/lib/ui/components/chips/chips_choice.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/ods_flutter_app_localizations.dart';
-import 'package:flutter_svg/svg.dart';
 import 'package:ods_flutter/components/app_bar/top/ods_top_app_bars.dart';
-import 'package:ods_flutter/components/chips/ods_filter_chips.dart';
-import 'package:ods_flutter/components/chips/ods_input_chips.dart';
+import 'package:ods_flutter/components/chips/ods_chip_common.dart';
+import 'package:ods_flutter/components/chips/ods_choice_chips.dart';
 import 'package:ods_flutter/components/lists/ods_list_switch.dart';
 import 'package:ods_flutter/components/sheets_bottom/ods_sheets_bottom.dart';
 import 'package:ods_flutter/guidelines/spacings.dart';
@@ -12,13 +11,13 @@ import 'package:ods_flutter_demo/ui/components/chips/chips_customization.dart';
 import 'package:ods_flutter_demo/ui/components/chips/chips_enum.dart';
 import 'package:ods_flutter_demo/ui/theme/theme_selector.dart';
 
-class ComponentChipsInput extends StatefulWidget {
-  const ComponentChipsInput({super.key});
+class ComponentChipsChoice extends StatefulWidget {
+  const ComponentChipsChoice({super.key});
   @override
-  State<ComponentChipsInput> createState() => _ComponentChipsInputState();
+  State<ComponentChipsChoice> createState() => _ComponentChipsChoiceState();
 }
 
-class _ComponentChipsInputState extends State<ComponentChipsInput> {
+class _ComponentChipsChoiceState extends State<ComponentChipsChoice> {
   bool isFiltered = true;
 
   @override
@@ -35,7 +34,7 @@ class _ComponentChipsInputState extends State<ComponentChipsInput> {
           title: AppLocalizations.of(context)!.componentCustomizeTitle,
         ),
         appBar: OdsAppTopBars(
-            title: AppLocalizations.of(context)!.componentChipInput,
+            title: AppLocalizations.of(context)!.componentChipChoice,
             actions: [ThemeSelector()],
             leading: BackButton()),
         body: SafeArea(child: _Body()),
@@ -57,8 +56,6 @@ class _BodyState extends State<_Body> {
   Widget build(BuildContext context) {
     final ChipsCustomizationState? customizationState =
         ChipsCustomization.of(context);
-    var colorScheme = Theme.of(context).colorScheme;
-    var colorFilter = ColorFilter.mode(colorScheme.secondary, BlendMode.srcIn);
 
     return Padding(
       padding:
@@ -66,32 +63,40 @@ class _BodyState extends State<_Body> {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          Text(AppLocalizations.of(context)!.componentChipFilterDescription,
+          Text(AppLocalizations.of(context)!.componentChipChoiceDescription,
               style: Theme.of(context).textTheme.bodyMedium),
           SizedBox(height: spacingM),
           Wrap(
             spacing: spacingS,
-            children: List<Widget>.generate(1, (int index) {
-              return OdsInputChip(
-                text: OdsApplication.recipes[index].title,
-                leadingAvatar:
-                    customizationState?.selectedElement == ChipsEnum.avatar
-                        ? CircleAvatar(
-                            backgroundImage: NetworkImage(
-                              OdsApplication.recipes[index].url,
-                            ),
-                          )
-                        : null,
-                leadingIcon:
-                    customizationState?.selectedElement == ChipsEnum.icon
-                        ? SvgPicture.asset("assets/recipes/ic_cooking_pot.svg",
-                            colorFilter: colorFilter)
-                        : null,
-                onClick: customizationState?.hasEnabled == true ? () {} : null,
-                onCancel: customizationState?.hasEnabled == true ? () {} : null,
-                enabled: customizationState?.hasEnabled == false ? false : true,
-              );
-            }),
+            children: List<Widget>.generate(
+              4,
+              (int index) {
+                bool isSelected = selectedIndex == index;
+                OdsChipLeadingAvatar? avatar;
+                if (customizationState?.selectedElement == ChipsEnum.none) {
+                  avatar = null;
+                } else if (customizationState?.selectedElement ==
+                    ChipsEnum.avatar) {
+                  avatar = OdsChipLeadingAvatar(
+                    image: NetworkImage(OdsApplication.foods[index + 41].image),
+                  );
+                }
+
+                return OdsChoiceChip(
+                  text: OdsApplication.foods[index + 41].name,
+                  onClick: (selected) {
+                    setState(() {
+                      selectedIndex = index;
+                      isFiltered = selected!;
+                    });
+                  },
+                  leadingAvatar: avatar,
+                  selected: isSelected,
+                  enabled:
+                      customizationState?.hasEnabled == false ? false : true,
+                );
+              },
+            ).toList(),
           ),
         ],
       ),
@@ -131,14 +136,14 @@ class _CustomizationContentState extends State<_CustomizationContent> {
             child: SingleChildScrollView(
               scrollDirection: Axis.horizontal,
               child: Row(
-                children: List<Widget>.generate(3, (int index) {
+                children: List<Widget>.generate(2, (int index) {
                   ChipsEnum? currentElement =
                       customizationState?.elements[index];
                   bool isSelected =
                       currentElement == customizationState?.selectedElement;
                   return Padding(
                     padding: EdgeInsets.only(right: 5, left: 10),
-                    child: OdsFilterChip(
+                    child: OdsChoiceChip(
                       text: customizationState?.elements[index]
                               .stringValue(context) ??
                           '',