+++ app/lib/ui/components/lists/lists_selection.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/ods_flutter_app_localizations.dart';
-import 'package:ods_flutter/components/chips/ods_filter_chips.dart';
+import 'package:ods_flutter/components/chips/ods_choice_chips.dart';
 import 'package:ods_flutter/components/lists/ods_list_selection_item.dart';
 import 'package:ods_flutter/components/lists/ods_list_switch.dart';
 import 'package:ods_flutter/components/sheets_bottom/ods_sheets_bottom.dart';
@@ -166,11 +166,10 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                       customizationState.selectedLeadingElement;
                   return Padding(
                     padding: EdgeInsets.only(right: spacingXs, left: spacingS),
-                    child: OdsFilterChips(
-                      label: customizationState.leadingElements[index]
+                    child: OdsChoiceChip(
+                      text: customizationState.leadingElements[index]
                           .stringValue(context),
-                      isSelected: isSelected,
-                      onSelected: (selected) {
+                      onClick: (selected) {
                         setState(
                           () {
                             selectedLeadingIndex = index;
@@ -180,6 +179,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                           },
                         );
                       },
+                      selected: isSelected,
                     ),
                   );
                 }),
@@ -210,11 +210,10 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                       customizationState.selectedTrailingElement;
                   return Padding(
                     padding: EdgeInsets.only(right: spacingXs, left: spacingS),
-                    child: OdsFilterChips(
-                      label: customizationState.trailingElements[index]
+                    child: OdsChoiceChip(
+                      text: customizationState.trailingElements[index]
                           .stringValue(context),
-                      isSelected: isSelected,
-                      onSelected: (selected) {
+                      onClick: (selected) {
                         setState(
                           () {
                             selectedTrailingIndex = index;
@@ -224,6 +223,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                           },
                         );
                       },
+                      selected: isSelected,
                     ),
                   );
                 }),