+++ app/lib/ui/components/progress/progress_linear.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/ods_flutter_app_localizations.dart';
 import 'package:ods_flutter/components/app_bar/top/ods_top_app_bars.dart';
-import 'package:ods_flutter/components/chips/ods_filter_chips.dart';
+import 'package:ods_flutter/components/chips/ods_choice_chips.dart';
 import 'package:ods_flutter/components/lists/ods_list_switch.dart';
 import 'package:ods_flutter/components/progress/ods_linear_progress_indicator.dart';
 import 'package:ods_flutter/components/sheets_bottom/ods_sheets_bottom.dart';
@@ -112,11 +112,10 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                     currentElement == customizationState.selectedProgressType;
                 return Padding(
                   padding: EdgeInsets.only(right: spacingXs, left: spacingS),
-                  child: OdsFilterChips(
-                    label: customizationState.progressType[index]
+                  child: OdsChoiceChip(
+                    text: customizationState.progressType[index]
                         .stringValue(context),
-                    isSelected: isSelected,
-                    onSelected: (selected) {
+                    onClick: (selected) {
                       setState(
                         () {
                           selectedLeadingIndex = index;
@@ -131,6 +130,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                         },
                       );
                     },
+                    selected: isSelected,
                   ),
                 );
               }),